requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
grad-cam>=1.4.0
streamlit 